requirements.txt
PyPDF2==3.0.1
pdfplumber==0.9.0
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
Pillow==10.0.0